moved styles for forms into forms.module.css instead of them being in listView.module.css
