+++ frontend/src/components/elements/ListViewElement/InsertFormElements/RoutesInsertFormComponent.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {ClickAwayListener} from '@mui/base';
-import styles from '@/styles/listView.module.css';
+import styles from '@/styles/forms.module.css';
 import {Button, FormControl, InputLabel, MenuItem, Select, SelectChangeEvent, TextField} from '@mui/material';
 import {LocationGroup, ScheduleDefinition} from '@/api/models';
 import {postScheduleDefinition} from '@/api/api';
